requirements.txt
flask
flask-httpauth
flask-restful
flask-sqlalchemy